feat: Reorganize custom widgets
