+++ lib/screens/events_browser/filter.dart
@@ -19,7 +19,7 @@
 
 import 'package:bluecherry_client/providers/server_provider.dart';
 import 'package:bluecherry_client/utils/extensions.dart';
-import 'package:bluecherry_client/utils/widgets/tree_view.dart';
+import 'package:bluecherry_client/widgets/tree_view.dart';
 import 'package:bluecherry_client/screens/events_browser/events_screen.dart';
 import 'package:bluecherry_client/widgets/misc.dart';
 import 'package:bluecherry_client/widgets/search.dart';
