+++ lib/screens/layouts/desktop/device_info_dialog.dart
@@ -19,7 +19,7 @@
 
 import 'package:bluecherry_client/models/device.dart';
 import 'package:bluecherry_client/utils/theme.dart';
-import 'package:bluecherry_client/utils/widgets/squared_icon_button.dart';
+import 'package:bluecherry_client/widgets/squared_icon_button.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
