+++ lib/screens/servers/wizard.dart
@@ -23,7 +23,8 @@ import 'package:bluecherry_client/screens/servers/add_server_info.dart';
 import 'package:bluecherry_client/screens/servers/additional_server_settings.dart';
 import 'package:bluecherry_client/screens/servers/configure_dvr_server.dart';
 import 'package:bluecherry_client/screens/servers/finish.dart';
-import 'package:bluecherry_client/utils/widgets/squared_icon_button.dart';
+import 'package:bluecherry_client/widgets/drawer_button.dart';
+import 'package:bluecherry_client/widgets/squared_icon_button.dart';
 import 'package:bluecherry_client/widgets/misc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
