+++ lib/screens/settings/mobile/settings.dart
@@ -28,7 +28,8 @@ import 'package:bluecherry_client/screens/servers/edit_server.dart';
 import 'package:bluecherry_client/screens/servers/edit_server_settings.dart';
 import 'package:bluecherry_client/utils/constants.dart';
 import 'package:bluecherry_client/utils/extensions.dart';
-import 'package:bluecherry_client/utils/widgets/squared_icon_button.dart';
+import 'package:bluecherry_client/widgets/drawer_button.dart';
+import 'package:bluecherry_client/widgets/squared_icon_button.dart';
 import 'package:bluecherry_client/screens/layouts/device_grid.dart';
 import 'package:bluecherry_client/widgets/misc.dart';
 import 'package:bluecherry_client/screens/settings/shared/date_language.dart';
