+++ lib/widgets/squared_icon_button.dart
